# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wholesome Uganda API Docs",
    "version": "1.0.0",
    "description": "API documentation for Wholesome Uganda website"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev server"
    },
    {
      "url": "https://api.wholesomeug.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/blogs": {
      "get": {
        "summary": "Get all blog posts",
        "tags": [
          "Blog"
        ],
        "responses": {
          "200": {
            "description": "List of blogs"
          }
        }
      },
      "post": {
        "summary": "Create a new blog post",
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "Blog date (YYYY-MM-DD)"
                  },
                  "category": {
                    "type": "string"
                  },
                  "excerpt": {
                    "type": "string"
                  },
                  "readTime": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog created"
          }
        }
      }
    },
    "/blogs/{id}": {
      "get": {
        "summary": "Get a blog post by ID",
        "tags": [
          "Blog"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog details"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update a blog post",
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "excerpt": {
                    "type": "string"
                  },
                  "readTime": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a blog post",
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/contact": {
      "post": {
        "summary": "Submit a contact message",
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Question about programs"
                  },
                  "message": {
                    "type": "string",
                    "example": "Tell me more about your initiatives."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message submitted"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "summary": "Get all contact messages",
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Messages per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of contact messages"
          }
        }
      }
    },
    "/contact/{id}": {
      "get": {
        "summary": "Get a single contact message",
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact message"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update contact message status",
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": {
                    "type": "string",
                    "example": "Updated subject"
                  },
                  "message": {
                    "type": "string",
                    "example": "Updated message body"
                  },
                  "status": {
                    "type": "string",
                    "example": "resolved"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated message"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact message",
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}